// upload/aetheria-terraforge-3d-extracted/src/types/game.ts
export type TerraformTool =
  | 'inspect'
  | 'elevate'   // +1 height
  | 'lower'     // -1 height
  | 'chasm'     // dig to 0
  | 'channel_lava'
  | 'channel_water'
  | 'flatten'
  | 'demolish';